requirements.txt
Flask==3.1.1
flask-cors==6.0.0
python-dotenv==1.0.1
requests==2.32.3
